make default setup work again
