+++ packages/graphql-app/components/product/product.tsx
@@ -1,20 +1,11 @@
 import React from 'react';
-import { useRouter } from 'next/router';
-import { useQuery } from '@apollo/react-hooks';
-import { GET_PRODUCT } from './get-product.graphql';
-import { ProductDetails } from '../../elements/product-details/product-details'
+import { ProductDetails } from '../../elements/product-details/product-details';
+import { useGetProductQuery } from '../../lib/_generated-types';
 
-export const Product = () => {
-    const router = useRouter();
-    const { id } = router.query;
-    console.log('render')
-    const { data } = useQuery(
-        GET_PRODUCT,
-        {
-          variables: { id: id.toString() },
-        }
-    );
+export const Product = ({ id }: any) => {
+    const { data } = useGetProductQuery({
+        variables: { id: id.toString() },
+    });
 
     return data ? <ProductDetails data={data} /> : null;
-           
 };
