+++ packages/graphql-app/components/products-list/get-products.graphql.ts
@@ -1,4 +1,4 @@
-import gql from 'graphql-tag'
+import gql from 'graphql-tag';
 
 export const GET_PRODUCTS = gql`
     query getProducts($id: String!) {
@@ -19,4 +19,4 @@ export const GET_PRODUCTS = gql`
             }
         }
     }
-`
+`;
