+++ packages/graphql-app/elements/product-details/product-details.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import {  GetProductQuery} from '../../lib/_generated-types';
+import { GetProductQuery } from '../../lib/_generated-types';
 import './product.scss';
 
 interface Props {
@@ -7,11 +7,11 @@ interface Props {
 }
 
 export const ProductDetails = ({ data }: Props) => {
-
     const image =
-    data && data.getProduct &&
-    data.getProduct.images &&
-    data.getProduct.images.filter((i: any) => i.key === 'L')[0].url;
+        data &&
+        data.getProduct &&
+        data.getProduct.images &&
+        data.getProduct.images.filter((i: any) => i.key === 'L')[0].url;
 
     return data && data.getProduct ? (
         <section>
@@ -21,14 +21,10 @@ export const ProductDetails = ({ data }: Props) => {
             </h1>
             <p>{data.getProduct.shortDescription}</p>
             <p>
-                <a
-                    href={data.getProduct.urls[0].value}
-                    target="_new"
-                >
+                <a href={data.getProduct.urls[0].value} target="_new">
                     see @bol.com
                 </a>
             </p>
         </section>
     ) : null;
-           
 };
