+++ packages/graphql-app/pages/product/[id].js
@@ -1,18 +1,20 @@
 import * as React from 'react';
 import { App } from '../../components/App';
 import { Product } from '../../components/product/product';
+import { withApollo } from '../../lib/apollo';
+import { useRouter } from 'next/router';
 
-export default class extends React.Component {
-    static async getInitialProps({ query }) {
-        return { query };
-    }
-    render() {
-        return (
-            <App>
-                <div className="content__section">
-                    <Product id={this.props.query.id} />
-                </div>
-            </App>
-        );
-    }
-}
+const ProductPage = () => {
+    const router = useRouter();
+    const { id } = router.query;
+
+    return (
+        <App>
+            <div className="content__section">
+                <Product id={id} />
+            </div>
+        </App>
+    );
+};
+
+export default withApollo(ProductPage);
